Refactor Account component: Extract user selection and freela publishing into separate components and fixed bug that caused the description field to always return null.
